requirements.txt
pandas>=2.0
numpy>=1.24
sentence-transformers>=2.2
faiss-cpu>=1.7
streamlit>=1.30
requests>=2.28
